Split runpath_list testing.
